Decompose Figbird query store
